Code refactoring, file management
