+++ templates/template.yml
@@ -1,7 +1,7 @@
 cv:
   personal_info:
     name: "Your Name"
-    photo_url: "./template/template.jpg"
+    photo_url: "../template/template.jpg"
     job: "Embedded Software Engineer"
     email: "your.email@example.com"
     phone_number: "06 xx xx xx xx"
